Add, modify and use jquery plugin to open modal on download
Each thumb is provided with a download button and by clicking on it we
want the user to see a message about future citations and then download
either the PDF or the SVG file. This has been done using the egrModal
plugin, which however was modified to make it possible to have different
modals on the same page. We need this, since each thumb will need a
different file to be downloaded when clicking on the buttons.